modularize the dashboard layouts into pages -first page is agent network -second page is ML experiments
